# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Label Computation System API
  version: 1.0.0
  description: |
    High-performance label computation system for quantitative trading that processes support/resistance level events 
    and computes forward-looking labels for pattern mining and ML models.
    
    Key features:
    - Dual-mode processing: Batch backfill (minutes) and real-time incremental (<1 second)
    - Multi-timeframe alignment using lower granularity data for path-dependent calculations
    - Support for 29 FX pairs and indices with billions of labels
    
    Critical SLAs:
    - Incremental compute: <100ms p99 latency
    - Batch throughput: 1M+ candles/minute
    - Cache hit rate: >95% for recent 24 hours
  contact:
    name: Label Computation System Team
    email: support@labelcompute.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.labelcompute.com/v1
    description: Production server
  - url: https://staging-api.labelcompute.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Development server
security:
  - bearerAuth: []
  - apiKey: []
...
paths:
...
  /labels/compute:
    post:
      tags: [Labels]
      summary: Compute labels for a single candle
      description: |
        Compute all configured labels for a single candle in real-time.
        Target latency: <100ms p99.
        
        For path-dependent labels, uses lower granularity data:
        - H4 → H1 data for horizon checks
        - D → H4 data for horizon checks  
        - W → D data for horizon checks
      operationId: computeLabels
      security:
        - bearerAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Unique request identifier for tracing
          required: false
          schema:
            type: string
            format: uuid
        - name: X-Cache-Strategy
          in: header
          description: Cache strategy hint
          required: false
          schema:
            type: string
            enum: [prefer_cache, bypass_cache, refresh_cache]
            default: prefer_cache
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CandleLabelRequest'
            example:
              instrument_id: "EURUSD"
              granularity: "H4"
              candle:
                ts: "2024-01-10T13:00:00Z"
                open: 1.0950
                high: 1.0970
                low: 1.0940
                close: 1.0965
                volume: 1250000
                atr_14: 0.0025
              label_types:
                - "enhanced_triple_barrier"
                - "vol_scaled_return"
                - "mfe_mae"
              options:
                horizon_periods: 6
                use_cache: true
      responses:
        '200':
          description: Labels computed successfully
          headers:
            X-Compute-Time-Ms:
              description: Computation time in milliseconds
              schema:
                type: integer
                example: 45
            X-Cache-Hit:
              description: Whether result came from cache
              schema:
                type: boolean
                example: false
            X-Rate-Limit-Remaining:
              description: Requests remaining in current window
              schema:
                type: integer
                example: 99
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComputedLabels'
              example:
                instrument_id: "EURUSD"
                granularity: "H4"
                ts: "2024-01-10T13:00:00Z"
                labels:
                  enhanced_triple_barrier:
                    label: 1
                    barrier_hit: "upper"
                    time_to_barrier: 3
                    barrier_price: 1.0990
                    level_adjusted: true
                  vol_scaled_return:
                    value: 2.35
                    quantile: 0.85
                  mfe_mae:
                    mfe: 0.0045
                    mae: -0.0012
                    profit_factor: 3.75
                computation_time_ms: 45
                cache_hit: false
                version: "1.0.0"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalError'

  # ===========================================================================
  # BATCH OPERATIONS
  # ===========================================================================
...
  /labels:
    get:
      tags: [Labels]
      summary: Query computed labels
      description: |
        Retrieve computed labels with flexible filtering and pagination.
        Optimized for time-series queries with proper indexing.
      operationId: getLabels
      security:
        - bearerAuth: []
      parameters:
        - name: instrument_id
          in: query
          required: true
          description: Instrument identifier
          schema:
            type: string
          example: "EURUSD"
        - name: granularity
          in: query
          required: true
          description: Time granularity
          schema:
            type: string
            enum: [M15, H1, H4, D, W]
          example: "H4"
        - name: start_date
          in: query
          required: true
          description: Start date (inclusive)
          schema:
            type: string
            format: date-time
          example: "2024-01-01T00:00:00Z"
        - name: end_date
          in: query
          required: true
          description: End date (exclusive)
          schema:
            type: string
            format: date-time
          example: "2024-01-31T23:59:59Z"
        - name: label_types
          in: query
          description: Comma-separated list of label types to include
          required: false
          schema:
            type: string
          example: "enhanced_triple_barrier,vol_scaled_return"
        - name: enhanced_triple_barrier_label
          in: query
          description: Filter by enhanced triple barrier outcome
          required: false
          schema:
            type: integer
            enum: [-1, 0, 1]
        - name: return_quantile_min
          in: query
          description: Minimum return quantile (0-100)
          required: false
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - name: return_quantile_max
          in: query
          description: Maximum return quantile (0-100)
          required: false
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - name: page
          in: query
          description: Page number (1-based)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          description: Items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: sort
          in: query
          description: Sort field and direction
          required: false
          schema:
            type: string
            enum: [ts_asc, ts_desc, forward_return_asc, forward_return_desc]
            default: ts_asc
      responses:
        '200':
          description: Labels retrieved successfully
          headers:
            X-Query-Time-Ms:
              description: Query execution time in milliseconds
              schema:
                type: integer
                example: 125
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelsList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
  /labels/{instrument_id}/{granularity}/{timestamp}:
    get:
      tags: [Labels]
      summary: Get labels for specific candle
      description: Retrieve all labels for a specific candle timestamp
      operationId: getLabelsByTimestamp
      security:
        - bearerAuth: []
      parameters:
        - name: instrument_id
          in: path
          required: true
          description: Instrument identifier
          schema:
            type: string
          example: "EURUSD"
        - name: granularity
          in: path
          required: true
          description: Time granularity
          schema:
            type: string
            enum: [M15, H1, H4, D, W]
          example: "H4"
        - name: timestamp
          in: path
          required: true
          description: Candle timestamp (ISO 8601)
          schema:
            type: string
            format: date-time
          example: "2024-01-10T13:00:00Z"
        - name: label_types
          in: query
          description: Comma-separated list of label types to include
          required: false
          schema:
            type: string
          example: "enhanced_triple_barrier,vol_scaled_return"
      responses:
        '200':
          description: Labels retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComputedLabels'
        '404':
          $ref: '#/components/responses/NotFound'

  # ===========================================================================
  # HEALTH & MONITORING
  # ===========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token containing user claims and permissions.
        Header format: `Authorization: Bearer <token>`
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for service-to-service authentication.
        Header format: `X-API-Key: <key>`
  schemas:
...
    CandleLabelRequest:
      type: object
      required: [instrument_id, granularity, candle]
      properties:
        instrument_id:
          type: string
          description: Instrument identifier (e.g., EURUSD, GBPJPY)
          pattern: '^[A-Z]{6}$|^[A-Z0-9]+$'
          example: "EURUSD"
        granularity:
          type: string
          enum: [M15, H1, H4, D, W]
          description: Time granularity for labels
        candle:
          $ref: '#/components/schemas/CandleData'
        label_types:
          type: array
          items:
            type: string
          description: Specific label types to compute (optional, computes all if empty)
          example: ["enhanced_triple_barrier", "vol_scaled_return"]
        options:
          type: object
          properties:
            horizon_periods:
              type: integer
              minimum: 1
              maximum: 100
              default: 6
              description: Forward-looking horizon in periods
            use_cache:
              type: boolean
              default: true
              description: Whether to use cached results
            force_recompute:
              type: boolean
              default: false
              description: Force recomputation even if cached
...
    ComputedLabels:
      type: object
      required: [instrument_id, granularity, ts, labels, computation_time_ms, version]
      properties:
        instrument_id:
          type: string
        granularity:
          type: string
          enum: [M15, H1, H4, D, W]
        ts:
          type: string
          format: date-time
        labels:
          type: object
          description: Computed label values
          properties:
            enhanced_triple_barrier:
              $ref: '#/components/schemas/EnhancedTripleBarrierLabel'
            vol_scaled_return:
              $ref: '#/components/schemas/VolScaledReturnLabel'
            mfe_mae:
              $ref: '#/components/schemas/MfeMaeLabel'
            return_quantile:
              type: integer
              minimum: 0
              maximum: 100
              description: Return quantile bucket (0-100)
            forward_return:
              type: number
              format: double
              description: Forward return over horizon
            # Additional labels would be defined here
        computation_time_ms:
          type: integer
          minimum: 0
          description: Time taken to compute labels
        cache_hit:
          type: boolean
          description: Whether result came from cache
        version:
          type: string
          description: Label computation version
          example: "1.0.0"
...
    LabelsList:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ComputedLabels'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
...
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "BAD_REQUEST"
              message: "Invalid request parameters"
              trace_id: "req_abc123def456"
    Unauthorized:
      description: Authentication required
      headers:
        WWW-Authenticate:
          description: Authentication method required
          schema:
            type: string
            example: 'Bearer realm="api", error="invalid_token"'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Authentication required"
              trace_id: "req_abc123def456"
    ValidationError:
      description: Validation error - invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "Invalid input data"
              details:
                - field: "candle.close"
                  message: "Must be greater than 0"
                  code: "MIN_VALUE"
                - field: "granularity"
                  message: "Must be one of: M15, H1, H4, D, W"
                  code: "INVALID_ENUM"
              trace_id: "req_abc123def456"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "NOT_FOUND"
              message: "Labels not found for specified parameters"
              trace_id: "req_abc123def456"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "INTERNAL_ERROR"
              message: "An internal error occurred"
              trace_id: "req_abc123def456"